Changer type des methodes valeur et mise a jour les test

+++ Addition.java
@@ -4,7 +4,7 @@ public class Addition extends Operation {
         super(operande1, operande2);
     }
 
-    public int valeur()
+    public double valeur()
     {
         return getOperande1().valeur()+getOperande2().valeur();
     }
